requirements.txt
Django==2.0.7
django-rest-multiple-models==2.1.1
djangorestframework==3.8.2
pytz==2018.5
